Consolidated true_color.py verification scripts. Modified custom_logger to use non underline as default
